refactor: collapse barrel shims and tidy remaining loose ends
src/lib/subnetCalculator.ts and src/lib/entraIdRbacService.ts existed only
to preserve import paths after their implementations moved into
src/lib/subnet/ and src/lib/entraId/. Nothing outside this repo imports
them, so point the call sites at the real barrels and delete the shims.

Within the subnet library, toUint32 was defined twice and subnetNetmask
reimplemented the bit arithmetic it documents itself as aliasing; it now
calls prefixToMask (verified identical for prefixes 0-32).

Smaller items:
- gray-matter is only used by build scripts, so move it to devDependencies
- index.tsx destructured the generated site data at module scope, which
  prevents Next from stripping it from the client bundle; every other page
  keeps the reference inside getStaticProps, so do the same here
- apply the spreadsheet-injection sanitiser to the RBAC and Entra ID CSV
  exporters, which skipped it while the IP exporter applied it
- load RegionMap via next/dynamic so d3-geo and topojson stay out of the
  ip-lookup entry bundle until a result actually has a region
- drop the "Update Entra ID roles data" step from the data workflow: its
  output is gitignored, so `git add -A` discarded it, and the deploy build
  already fetches the file via fetchEntraIdRolesIfPossible
- correct the Entra ID empty-state text, which claimed the workflow commits
  entraid-roles.json
- validateData.js swallowed unexpected errors and let the build proceed
