+++ package.json
@@ -29,7 +29,6 @@
     "@types/topojson-client": "^3.1.5",
     "d3-geo": "^3.1.1",
     "exceljs": "^4.4.0",
-    "gray-matter": "^4.0.3",
     "ip-cidr": "^4.0.2",
     "next": "^16.3.0",
     "papaparse": "^5.5.4",
@@ -60,6 +59,7 @@
     "@types/topojson-specification": "^1.0.5",
     "eslint": "^10.8.1",
     "eslint-plugin-unused-imports": "^4.4.1",
+    "gray-matter": "^4.0.3",
     "postcss": "^8.5.26",
     "tailwindcss": "^4.3.3",
     "ts-node": "^10.9.2",
